pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autotrans"
version = "0.1.0"
description = "Realtime desktop OCR translation overlay for games on Windows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.8",
  "numpy>=1.26",
  "Pillow>=10.4",
  "mss>=10.0",
  "rapidfuzz>=3.9",
  "httpx>=0.28",
  "pydantic>=2.9",
  "pywin32>=308",
]

[project.optional-dependencies]
ocr = [
  "paddleocr>=2.8",
  "opencv-python-headless>=4.10",
  "paddlepaddle>=3.3",
  "rapidocr_onnxruntime>=1.2",
]
local_translate = ["ctranslate2>=4.5", "sentencepiece>=0.2", "huggingface_hub>=0.36"]
openai = ["openai>=1.68"]
dev = ["pytest>=8.3", "pytest-qt>=4.4"]

[project.scripts]
autotrans = "autotrans.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]